Implemented solutions for version1
