+++ workspace-version1/CategorySrv/src/main/java/webshop/categories/api/BaseResponse.java
@@ -2,9 +2,9 @@ package webshop.categories.api;
 
 import javax.validation.constraints.NotNull;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-
 import org.hibernate.validator.constraints.NotEmpty;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class BaseResponse {
     @NotNull
