+++ workspace-version1/CategorySrv/src/main/java/webshop/categories/api/ProductCategory.java
@@ -3,6 +3,7 @@ package webshop.categories.api;
 import java.util.List;
 
 import javax.validation.constraints.Min;
+
 import org.hibernate.validator.constraints.NotEmpty;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
