+++ workspace-version1/NotificationSrv/src/main/java/webshop/notifications/api/BaseResponse.java
@@ -2,9 +2,9 @@ package webshop.notifications.api;
 
 import javax.validation.constraints.NotNull;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-
 import org.hibernate.validator.constraints.NotEmpty;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class BaseResponse {
     @NotNull
