+++ workspace-version1/OrderSrv/src/main/java/webshop/orders/api/BaseResponse.java
@@ -2,9 +2,9 @@ package webshop.orders.api;
 
 import javax.validation.constraints.NotNull;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-
 import org.hibernate.validator.constraints.NotEmpty;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class BaseResponse {
     @NotNull
