+++ workspace-version1/OrderSrv/src/main/java/webshop/orders/api/CreditRatingCheckResponse.java
@@ -1,6 +1,7 @@
 package webshop.orders.api;
 
 import javax.validation.constraints.Min;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class CreditRatingCheckResponse {
