+++ workspace-version1/OrderSrv/src/main/java/webshop/orders/resources/OrderResource.java
@@ -20,15 +20,16 @@ import javax.ws.rs.client.Invocation;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response.Status;
 
-import com.codahale.metrics.annotation.Timed;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import com.codahale.metrics.annotation.Timed;
 
 import io.dropwizard.jersey.params.IntParam;
 import io.dropwizard.jersey.params.LongParam;
 import webshop.orders.api.BaseResponse;
 import webshop.orders.api.CreditRatingCheckResponse;
+import webshop.orders.api.MarketingMailRequest;
 import webshop.orders.api.Order;
 import webshop.orders.api.OrderItem;
 import webshop.orders.api.ProductAvailabilityCheckResponse;
@@ -41,8 +42,7 @@ public class OrderResource {
 	private Client restClient;
 	private Logger log;
 	private final String CREDIT_RATING_CHECK_ENDPOINT = "http://localhost:8000";
-	// TODO Ex2, Task3: Change to the URL of the new WarehouseSrv
-	private final String PRODUCT_AVAILABILITY_CHECK_ENDPOINT = "http://localhost:8050";
+	private final String PRODUCT_AVAILABILITY_CHECK_ENDPOINT = "http://localhost:8070";
 
 	public OrderResource(OrderRepository repository, Client restClient) {
 		this.orderRepository = repository;
@@ -115,9 +115,13 @@ public class OrderResource {
 				final Order createdOrder = orderRepository.store(order);
 				log.info("Order with ID " + createdOrder.getId() + " successfully created.");
 
-				// TODO Ex1, Task3: Invoke NotificationSrv to send marketing mail
-				
-				
+				// Invoking the NotificationSrv to send a SIMILAR_PRODUCTS_MAIL for the new
+				// order
+				MarketingMailRequest marketingMailRequest = new MarketingMailRequest("SIMILAR_PRODUCTS_MAIL",
+						createdOrder);
+				Invocation.Builder request = restClient.target("http://localhost:8010/marketing-mails").request();
+				request.post(Entity.json(marketingMailRequest), BaseResponse.class);
+
 				// Return final response
 				return new BaseResponse("OK", 201, "Order with ID " + createdOrder.getId() + " successfully created.");
