+++ workspace-version1/WarehouseSrv/src/main/java/webshop/warehouse/api/ProductAvailabilityCheckResponse.java
@@ -1,6 +1,7 @@
 package webshop.warehouse.api;
 
 import javax.validation.constraints.Min;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class ProductAvailabilityCheckResponse {
